requirements.txt
discord.py
dnspython
youtube_dl
ffmpeg
colorlog
PyNaCl
pillow